yt_downloader/back/yt_downloader/requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
gunicorn==23.0.0
typing_extensions==4.13.2
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
djangorestframework==3.16.0
yt-dlp==2025.5.22
